updates
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@763 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ modules/tibrv/QoreTibrvFtMember.h
@@ -143,7 +143,7 @@ inline QoreNode *QoreTibrvFtMember::getEvent(class ExceptionSink *xsink)
       if (status == TIBRV_INVALID_QUEUE)
       {
 	 class Hash *h = new Hash();
-	 h->setKeyValue("action", new QoreNode(-1LL), NULL);
+	 h->setKeyValue("action", new QoreNode((int64)-1), NULL);
 	 return new QoreNode(h);
       }
