+++ modules/tibrv/QoreTibrvFtMonitor.h
@@ -137,7 +137,7 @@ inline QoreNode *QoreTibrvFtMonitor::getEvent(class ExceptionSink *xsink)
       if (status == TIBRV_INVALID_QUEUE)
       {
 	 class Hash *h = new Hash();
-	 h->setKeyValue("action", new QoreNode(-1LL), NULL);
+	 h->setKeyValue("action", new QoreNode((int64)-1), NULL);
 	 return new QoreNode(h);
       }
